requirements.txt
flask==2.1.1
werkzeug==2.1.1
flask-socketio==5.1.1
flask-cors==3.0.10
python-socketio==5.4.0
spotipy==2.19.0
yt-dlp==2023.7.6
mutagen==1.45.1
requests==2.26.0
urllib3==1.26.15
charset-normalizer==2.0.4
chardet==4.0.0
Pillow==10.1.0
SQLAlchemy==1.4.46
firebase-admin==6.2.0
google-auth==2.23.0
google-auth-oauthlib==1.1.0
google-cloud-firestore==2.13.1
flask-login==0.6.3
python-dotenv==1.0.0
